% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_update.R
\name{check_update}
\alias{check_update}
\title{FUNCTION_TITLE}
\usage{
check_update(user, datetime = gittime(Sys.Date() - 1),
  gh_pat = Sys.getenv("GITHUB_PAT"), verbose = TRUE)
}
\arguments{
\item{user}{PARAM_DESCRIPTION}

\item{datetime}{PARAM_DESCRIPTION, Default: gittime(Sys.Date()-1)}

\item{gh_pat}{character, github pat Default: Sys.getenv('GITHUB_PAT')}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{add_headers}},\code{\link[httr]{stop_for_status}}
}
