% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot.R
\name{create_plot}
\alias{create_plot}
\title{create network plot}
\usage{
create_plot(data, add_labels = "none")
}
\arguments{
\item{data}{tibble, tidied commit data from \code{\link{get_commit}}}

\item{add_labels}{character, add labels to graph nodes, Default: 'none'}
}
\value{
ggraph plot
}
\description{
creates a ggpraph object using the tidied commit data
}
\details{
add_labels can be c('both','repo','user','none')
}
\examples{

\dontrun{
gh_data_2018 <- gh_data\%>\%
dplyr::filter(format(date,'\%Y')=='2018')\%>\%
dplyr::mutate(freq_var = format(date,'\%Y'))\%>\%
create_graph_data()

gh_data_2018\%>\%create_plot()

gh_data_2018\%>\%create_plot(add_labels='user')

gh_data_2018\%>\%create_plot(add_labels='repo')

gh_data_2018\%>\%create_plot(add_labels='both')

}

}
\seealso{
\code{\link[tidygraph]{as_tbl_graph.data.frame}},\code{\link[tidygraph]{centrality}}
}
