% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot.R
\name{create_graph_data}
\alias{create_graph_data}
\title{create data for network plot}
\usage{
create_graph_data(data, touch = NULL)
}
\arguments{
\item{data}{data.frame, tidied commit data from \code{\link{get_commit}}}

\item{touch}{highlight a user, Default: NULL}
}
\value{
tibble
}
\description{
creates tidied commit data to pass into plot
}
\seealso{
\code{\link[tidygraph]{as_tbl_graph.data.frame}},\code{\link[tidygraph]{centrality}}
}
