% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_commit_data.R
\name{get_commit}
\alias{get_commit}
\title{Use github API to return tidy table of commits}
\usage{
get_commit(user, repo, gh_pat = NULL)
}
\arguments{
\item{user}{user name}

\item{repo}{repository name}

\item{gh_pat}{character, github pat Default: NULL}
}
\value{
tibble
}
\description{
Use github api rest call to fetch commit information
for a repository
}
\examples{

\dontrun{

gh_data <- get_commit(user = 'thinkr-open', repo = 'remedy')

}

}
\seealso{
\code{\link[httr]{GET}}
}
