% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rate_limit.R
\name{get_rate_limit}
\alias{get_rate_limit}
\title{Get current Github API rate limit}
\usage{
get_rate_limit(gh_pat = NULL)
}
\arguments{
\item{gh_pat}{character, github pat Default: NULL}
}
\value{
tibble
}
\description{
Get the current Github API rate limit with time to reset in minutes
}
\details{
using this will not count against the rate limit
}
\examples{

get_rate_limit()

get_rate_limit(gh_pat = Sys.getenv('GITHUB_PAT'))

}
