% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_commit_data.R
\name{gh_commit_data}
\alias{gh_commit_data}
\title{Use github API to return tidy table of commits}
\usage{
gh_commit_data(repo, gh_pat = NULL)
}
\arguments{
\item{repo}{repository name (user/repo)}

\item{gh_pat}{character, github pat Default: NULL}
}
\value{
dataframe
}
\description{
Use github api rest call to fetch commit information
for a repository
}
\details{
beware of the rate limit for the api
}
\examples{

\dontrun{
repos <- c(
'thinkr-open/remedy',
'hrbrmstr/slackr',
'hrbrmstr/ggalt',
'hrbrmstr/hrbrthemes',
'metrumresearchgroup/sinew')

#single

gh_data <- gh_commit_data(repos[1])


#multiple
gh_data <- purrr::map_df(repos,gh_commit_data)

}

}
\seealso{
\code{\link[jsonlite]{read_json}}
}
