% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_update.R
\name{gittime}
\alias{gittime}
\title{FUNCTION_TITLE}
\usage{
gittime(datetime = Sys.time())
}
\arguments{
\item{datetime}{PARAM_DESCRIPTION, Default: Sys.time()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
