% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tidyverse.R
\docType{data}
\name{tidyverse}
\alias{tidyverse}
\title{tidyverse commit data}
\format{A data frame with 397 rows and 7 variables:
\describe{
  \item{\code{w}}{double Start of the week in epoch time}
  \item{\code{a}}{double lines added}
  \item{\code{d}}{double lines deleted}
  \item{\code{c}}{double number of commits}
  \item{\code{user}}{character username}
  \item{\code{date}}{double w as.Date}
  \item{\code{repo}}{character repository name}
}}
\usage{
tidyverse
}
\description{
commit data from tidyverse github repository
}
\keyword{datasets}
