% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_commit_data.R
\name{get_commit}
\alias{get_commit}
\title{Use github API to return tidy table of commits}
\usage{
get_commit(owner, repo, gh_pat = Sys.getenv("GITHUB_PAT"))
}
\arguments{
\item{owner}{user name}

\item{repo}{repository name}

\item{gh_pat}{character, github pat, Default: Sys.getenv('GITHUB_PAT')}
}
\value{
tibble
}
\description{
Use github api rest call to fetch commit information
for a repository
}
\examples{

\dontrun{

gh_data <- get_commit(owner = 'thinkr-open', repo = 'remedy')

}

}
\seealso{
\code{\link[httr]{GET}}
}
