% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_plots.R
\name{gh_plots}
\alias{gh_plots}
\title{create network plots of the tidied commit data}
\usage{
gh_plots(plots, add_labels = "none", title = NULL)
}
\arguments{
\item{plots}{output from \code{\link{create_plot}}}

\item{title}{character title of plot, Default: NULL}

\item{add_label}{character, add labels to graph nodes, Default: 'none'}
}
\value{
plot
}
\description{
use ggraph to create logitudinal network plots by year
}
\details{
add_labels can be c('both','repo','user','none')
}
\examples{

gh_data\%>\%gh_plots(title = 'Github Repository Contribution Network')

gh_data\%>\%gh_plots(title = 'Github Repository Contribution Network',add_labels='both')

}
